#bugfix: trival

+++ application/hand.cpp
@@ -5,18 +5,18 @@
 class robotHand : public dst::configObject{
 private:
     void mouseDrag(const QJsonArray& aOrigin, const QJsonArray& aDelta){
-        SetCursorPos(aOrigin[0].toInt(), aOrigin[1].toInt());
+        SetCursorPos(aOrigin[0].toDouble(), aOrigin[1].toDouble());
         mouse_event(MOUSEEVENTF_LEFTDOWN, 0, 0, 0, 0);
-        mouse_event(MOUSEEVENTF_MOVE, aDelta[0].toInt(), aDelta[1].toInt(), 0, 0);
+        mouse_event(MOUSEEVENTF_MOVE, aDelta[0].toDouble(), aDelta[1].toDouble(), 0, 0);
         mouse_event(MOUSEEVENTF_LEFTUP, 0, 0, 0, 0);
     }
     void mouseClick(const QJsonArray& aOrigin){
-        SetCursorPos(aOrigin[0].toInt(), aOrigin[1].toInt());
+        SetCursorPos(aOrigin[0].toDouble(), aOrigin[1].toDouble());
         mouse_event(MOUSEEVENTF_LEFTDOWN, 0, 0, 0, 0);
         mouse_event(MOUSEEVENTF_LEFTUP, 0, 0, 0, 0);
     }
     void mouseMove(const QJsonArray& aOrigin){
-        SetCursorPos(aOrigin[0].toInt(), aOrigin[1].toInt());
+        SetCursorPos(aOrigin[0].toDouble(), aOrigin[1].toDouble());
     }
 public:
     robotHand() : configObject(QJsonObject()){
